Move "move up button" to layout to display on every part of page
